+++ lib/src/customization/header_style.dart
@@ -76,30 +76,32 @@ class HeaderStyle {
   /// Decoration of the header.
   final BoxDecoration decoration;
 
+  final Color selectMonthYearHighlightColor;
+
   /// Creates a `HeaderStyle` used by `WeekView Calendar` widget.
-  const HeaderStyle({
-    this.titleCentered = false,
-    this.formatButtonVisible = true,
-    this.formatButtonShowsNext = true,
-    this.titleTextFormatter,
-    this.titleTextStyle = const TextStyle(fontSize: 17.0),
-    this.formatButtonTextStyle = const TextStyle(fontSize: 14.0),
-    this.formatButtonDecoration = const BoxDecoration(
-      border: const Border.fromBorderSide(BorderSide()),
-      borderRadius: const BorderRadius.all(Radius.circular(12.0)),
-    ),
-    this.headerMargin = const EdgeInsets.all(0.0),
-    this.headerPadding = const EdgeInsets.symmetric(vertical: 8.0),
-    this.formatButtonPadding =
-        const EdgeInsets.symmetric(horizontal: 10.0, vertical: 4.0),
-    this.leftChevronPadding = const EdgeInsets.all(12.0),
-    this.rightChevronPadding = const EdgeInsets.all(12.0),
-    this.leftChevronMargin = const EdgeInsets.symmetric(horizontal: 8.0),
-    this.rightChevronMargin = const EdgeInsets.symmetric(horizontal: 8.0),
-    this.leftChevronIcon = const Icon(Icons.chevron_left),
-    this.rightChevronIcon = const Icon(Icons.chevron_right),
-    this.leftChevronVisible = true,
-    this.rightChevronVisible = true,
-    this.decoration = const BoxDecoration(),
-  });
+  const HeaderStyle(
+      {this.titleCentered = false,
+      this.formatButtonVisible = true,
+      this.formatButtonShowsNext = true,
+      this.titleTextFormatter,
+      this.titleTextStyle = const TextStyle(fontSize: 17.0),
+      this.formatButtonTextStyle = const TextStyle(fontSize: 14.0),
+      this.formatButtonDecoration = const BoxDecoration(
+        border: const Border.fromBorderSide(BorderSide()),
+        borderRadius: const BorderRadius.all(Radius.circular(12.0)),
+      ),
+      this.headerMargin = const EdgeInsets.all(0.0),
+      this.headerPadding = const EdgeInsets.symmetric(vertical: 8.0),
+      this.formatButtonPadding =
+          const EdgeInsets.symmetric(horizontal: 10.0, vertical: 4.0),
+      this.leftChevronPadding = const EdgeInsets.all(12.0),
+      this.rightChevronPadding = const EdgeInsets.all(12.0),
+      this.leftChevronMargin = const EdgeInsets.symmetric(horizontal: 8.0),
+      this.rightChevronMargin = const EdgeInsets.symmetric(horizontal: 8.0),
+      this.leftChevronIcon = const Icon(Icons.chevron_left),
+      this.rightChevronIcon = const Icon(Icons.chevron_right),
+      this.leftChevronVisible = true,
+      this.rightChevronVisible = true,
+      this.decoration = const BoxDecoration(),
+      this.selectMonthYearHighlightColor = const Color(0xFFBBDDFF)});
 }